Proto improvements
* Organizing loom.proto into separated files for: gateway, address
mapper and evm
* Improving proto and plasma abi generators
* Updated package.json
