+++ src/client.ts
@@ -2,8 +2,8 @@ import debug from 'debug'
 import { Message } from 'google-protobuf'
 import EventEmitter from 'events'
 import retry from 'retry'
+import { VMType } from './proto/loom_pb'
 import {
-  VMType,
   EvmTxReceipt,
   EvmTxObject,
   EthBlockInfo,
@@ -11,7 +11,7 @@ import {
   EthBlockHashList,
   EthFilterLogList,
   EthTxHashList
-} from './proto/loom_pb'
+} from './proto/evm_pb'
 import { Uint8ArrayToB64, B64ToUint8Array, bufferToProtobufBytes } from './crypto-utils'
 import { Address, LocalAddress } from './address'
 import { WSRPCClient, IJSONRPCEvent } from './internal/ws-rpc-client'
